+++ src/aspectlib/__init__.py
@@ -307,6 +307,7 @@ class Rollback(object):
     def __exit__(self, *_):
         for rollback in self._rollbacks:
             rollback()
+        self._rollbacks.clear()
 
     rollback = __call__ = __exit__