+++ CHANGELOG.rst
@@ -1,11 +1,12 @@
 Changelog
 =========
 
-
-1.3.4 (2016-04-08)
+1.4.0 (2016-04-09)
 ------------------
 
 * Corrected weaving of methods, the weaved function should be unbound.
+* Rolling back only applies undos once.
+* Added a convenience `weave` fixture for pytest.
 
 1.3.3 (2015-10-02)
 ------------------